projects/14-DeepDiary/Backend/requirements.txt
absl-py==0.13.0
addict==2.4.0
aiohttp==3.8.1
aiosignal==1.2.0
albumentations==1.1.0
aliyun-python-sdk-core==2.13.36
aliyun-python-sdk-kms==2.15.0
amqp==5.1.1
anyjson==0.3.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.4.1
asttokens==2.0.5
async-timeout==4.0.2
attrs==21.4.0
backcall==0.2.0
backports.zoneinfo==0.2.1
base58==2.1.1
beautifulsoup4==4.11.1
billiard==3.6.4.0
bitarray==2.6.0
bleach==5.0.0
boltons==21.0.0
cchardet==2.1.7
celery==5.2.7
celery-with-redis==3.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cmake==3.24.1
colorama==0.4.4
commonmark==0.9.1
coreschema==0.0.4
crcmod==1.7
cryptography==36.0.1
cycler==0.11.0
Cython==0.29.32
cytoolz==0.12.0
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
Django==3.2.9
django-allauth==0.47.0
django-appconf==1.0.5
django-ckeditor==6.2.0
django-cors-headers==3.13.0
django-extensions==3.1.5
django-filter==21.1
django-imagekit==4.1.0
django-js-asset==1.2.2
django-mdeditor==0.1.20
django-model-utils==4.2.0
django-mptt==0.13.4
django-notifications-hq==1.6.0
django-oss-storage==1.1.1
django-password-reset==2.0
django-redis==5.2.0
django-taggit==2.0.0
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.1.0
dnspython==2.2.1
easydict==1.9
entrypoints==0.4
eth-abi==2.2.0
eth-account==0.5.9
eth-hash==0.5.0
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.9.5
eventlet==0.33.1
executing==0.8.3
fastjsonschema==2.15.3
filestools==0.1.2
flatbuffers==2.0
fonttools==4.32.0
frozenlist==1.3.1
greenlet==1.1.3
hexbytes==0.3.0
idna==3.3
image==1.5.33
imageio==2.16.2
importlib-metadata==4.10.0
importlib-resources==5.7.1
insightface==0.6.2
interchange==2021.0.4
ipfshttpclient==0.8.0a2
ipykernel==6.13.0
ipython==8.3.0
ipython-genutils==0.2.0
jedi==0.18.1
Jinja2==3.1.2
jmespath==0.10.0
joblib==1.1.0
jsonfield==3.1.0
jsonschema==4.5.1
jupyter-client==7.3.0
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
kiwisolver==1.4.2
kombu==5.2.4
lru-dict==1.1.8
Markdown==3.3.6
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
model-index==0.1.11
monotonic==1.6
multiaddr==0.0.9
multidict==6.0.2
mysqlclient==2.1.0
nbclient==0.6.2
nbconvert==6.5.0
nbformat==5.4.0
neo4j-driver==4.4.3
nest-asyncio==1.5.5
netaddr==0.8.0
networkx==2.8
notebook==6.4.11
numpy==1.23.2
oauthlib==3.1.1
onnx==1.11.0
onnxruntime==1.11.0
opencv-contrib-python==4.5.5.64
opencv-python-headless==4.5.5.64
openmim==0.1.5
ordered-set==4.1.0
oss2==2.15.0
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
pansi==2020.7.3
parsimonious==0.8.1
parso==0.8.3
pickleshare==0.7.5
pilkit==2.0
Pillow==9.0.0
prettytable==3.2.0
prometheus-client==0.14.1
prompt-toolkit==3.0.29
protobuf==3.20.0
psutil==5.9.0
pure-eval==0.2.2
py2neo==2021.2.3
pycallgraph==1.0.1
pycocotools==2.0.4
pycparser==2.21
pycryptodome==3.14.1
pyexiv2==2.7.1
Pygments==2.11.2
PyJWT==1.7.1
pymongo==4.1.1
pyparsing==3.0.8
pyperclip==1.8.2
pyrsistent==0.18.1
python-dateutil==2.8.2
python-mcs-sdk==0.1.0
python3-openid==3.2.0
pytz==2021.3
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==22.3.0
q==2.6
qudida==0.0.4
redis==4.3.4
regex==2022.4.24
requests==2.27.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rich==11.0.0
rlp==2.0.1
scikit-image==0.19.2
scikit-learn==1.0.2
scipy==1.8.0
Send2Trash==1.8.0
signals==0.0.2
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
stack-data==0.2.0
swapper==1.3.0
tabulate==0.8.9
terminado==0.13.3
terminaltables==3.1.10
threadpoolctl==3.1.0
tifffile==2022.4.8
tinycss2==1.1.1
toolz==0.12.0
torch==1.12.1+cu113
torchaudio==0.12.1+cu113
torchvision==0.13.1+cu113
tornado==6.1
tqdm==4.64.0
traitlets==5.1.1
transaction==3.0.1
typing_extensions==4.1.1
tzdata==2021.5
urllib3==1.26.8
varint==1.0.2
vine==5.0.0
wcwidth==0.2.5
web3==5.30.0
webencodings==0.5.1
websockets==9.1
wrapt==1.14.1
yapf==0.32.0
yarl==1.8.1
zipp==3.7.0
zope.interface==5.4.0
